clear blog plan!
This is going to be great—I know basically what I want to write. I'm
going to want to rewrite the code to not bother with the bitvec
crate. Maybe as a standalone file to not even need the rand crate,
either? But maybe linking to the full repo (rather than a standalone
Gist) is OK?